+++ tests/integration/integrationTestHelper.js
@@ -18,10 +18,10 @@ class IntegrationTestHelper {
 		docDefinition.attachments = docDefinition.attachments || {};
 		var fontDescriptors = {
 			Roboto: {
-				normal: 'tests/fonts/Roboto-Regular.ttf',
-				bold: 'tests/fonts/Roboto-Medium.ttf',
-				italics: 'tests/fonts/Roboto-Italic.ttf',
-				bolditalics: 'tests/fonts/Roboto-Italic.ttf'
+				normal: 'fonts/Roboto/Roboto-Regular.ttf',
+				bold: 'fonts/Roboto/Roboto-Medium.ttf',
+				italics: 'fonts/Roboto/Roboto-Italic.ttf',
+				bolditalics: 'fonts/Roboto/Roboto-MediumItalic.ttf'
 			}
 		};