+++ tests/unit/Node-interface.spec.js
@@ -9,10 +9,10 @@ var pdfmake = require('../../js/index');
 
 pdfmake.addFonts({
 	Roboto: {
-		normal: 'tests/fonts/Roboto-Regular.ttf',
-		bold: 'tests/fonts/Roboto-Medium.ttf',
-		italics: 'tests/fonts/Roboto-Italic.ttf',
-		bolditalics: 'tests/fonts/Roboto-MediumItalic.ttf'
+		normal: 'fonts/Roboto/Roboto-Regular.ttf',
+		bold: 'fonts/Roboto/Roboto-Medium.ttf',
+		italics: 'fonts/Roboto/Roboto-Italic.ttf',
+		bolditalics: 'fonts/Roboto/Roboto-MediumItalic.ttf'
 	}
 });